+++ src/index.ts
@@ -1,4 +1,4 @@
 export { eslintConfigs, mergeEslintConfigPlugins } from './configs/eslint/index.js';
 export { eslintConfigFileTemplate } from './configs/eslint/template.js';
 export { installDevDependencies } from './setup/devDependencies.js';
-export type { ProjectType } from './types/index.js';
+export { addPackageJsonScripts } from './setup/packageJsonScripts.js';